// frontend/src/api/models/getTradePurchaseOrdersParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { GetTradePurchaseOrdersOrderDir } from './getTradePurchaseOrdersOrderDir'
import type { GetTradePurchaseOrdersStatus } from './getTradePurchaseOrdersStatus'

export type GetTradePurchaseOrdersParams = {
  /**
   * Search term (order number, supplier name)
   */
  search?: string
  /**
   * Supplier ID
   */
  supplier_id?: string
  /**
   * Warehouse ID
   */
  warehouse_id?: string
  /**
   * Order status
   */
  status?: GetTradePurchaseOrdersStatus
  /**
   * Multiple order statuses
   */
  statuses?: string[]
  /**
   * Start date (ISO 8601)
   */
  start_date?: string
  /**
   * End date (ISO 8601)
   */
  end_date?: string
  /**
   * Minimum payable amount
   */
  min_amount?: number
  /**
   * Maximum payable amount
   */
  max_amount?: number
  /**
   * Page number
   */
  page?: number
  /**
   * Page size
   * @maximum 100
   */
  page_size?: number
  /**
   * Order by field
   */
  order_by?: string
  /**
   * Order direction
   */
  order_dir?: GetTradePurchaseOrdersOrderDir
}