// frontend/src/api/models/getTradePurchaseOrdersStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type GetTradePurchaseOrdersStatus =
  (typeof GetTradePurchaseOrdersStatus)[keyof typeof GetTradePurchaseOrdersStatus]

export const GetTradePurchaseOrdersStatus = {
  draft: 'draft',
  confirmed: 'confirmed',
  partial_received: 'partial_received',
  completed: 'completed',
  cancelled: 'cancelled',
} as const